runtimes/python/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-multipart==0.0.6
python-jose==3.3.0
loguru==0.7.0
psycopg2-binary==2.9.5
psutil==5.9.5
